Type skyFn_80088c94 mode param u8 to mask bool arg (beginLoadingMap)
Caller now emits the clrlwi-to-u8 on the (x&N)?1:0 bool argument, matching
target (99.277->99.656). Remaining diff is score-neutral @144 relocs.

+++ src/main/shader.c
@@ -2526,7 +2526,7 @@ extern void clearSaveGameLoadingFlag(void);
 extern void mapSetupPlayer(void);
 extern void* saveGameGetEnvState(void);
 extern void getEnvfxAct(void* obj, void* source, int actId, int flags);
-extern void skyFn_80088c94(int flags, int mode);
+extern void skyFn_80088c94(int flags, u8 mode);
 extern void skyFn_80088e54(f32 a, int on);
 
 void beginLoadingMap(void)
